# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [{ "url": "http://localhost:8080", "description": "Generated server url" }],
  "paths": {
...
    "/api/plans/{id}": {
      "get": {
        "tags": ["plan-controller"],
        "operationId": "getPlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/PlanResponse" } } }
          }
        }
      },
      "put": {
        "tags": ["plan-controller"],
        "operationId": "updatePlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PlanRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/PlanResponse" } } }
          }
        }
      },
      "delete": {
        "tags": ["plan-controller"],
        "operationId": "deletePlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/plans": {
      "get": {
        "tags": ["plan-controller"],
        "operationId": "getAllPlans",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PlanResponse" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["plan-controller"],
        "operationId": "createPlan",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PlanRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "$ref": "#/components/schemas/PlanResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlanRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "monthlyPrice": { "type": "number", "minimum": 0.0 },
          "dataLimitGb": { "type": "integer", "format": "int32" },
          "minutesLimit": { "type": "integer", "format": "int32" },
          "smsLimit": { "type": "integer", "format": "int32" },
          "isActive": { "type": "boolean" }
        },
        "required": ["monthlyPrice", "name"]
      },
      "PlanResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "monthly_price": { "type": "number" },
          "data_limit_gb": { "type": "integer", "format": "int32" },
          "minutes_limit": { "type": "integer", "format": "int32" },
          "sms_limit": { "type": "integer", "format": "int32" },
          "is_active": { "type": "boolean" }
        }
      },
...
    }
  }
}